+++ src/drawable.hpp
@@ -13,13 +13,13 @@ public:
   inline static xcb_screen_t *screen_; ///< X screen
   int16_t x;                           ///< x coordinate of the top left corner
   int16_t y;                           ///< y coordinate of the top left corner
-  uint16_t width;
-  uint16_t height;
+  uint width;
+  uint height;
 
-  drawable (int16_t x,      ///< x coordinate of the top left corner
-            int16_t y,      ///< y coordinate of the top left corner
-            uint16_t width, ///< Width of the display
-            uint16_t height ///< Height of the display
+  drawable (int16_t x,  ///< x coordinate of the top left corner
+            int16_t y,  ///< y coordinate of the top left corner
+            uint width, ///< Width of the display
+            uint height ///< Height of the display
   );
 };
